updated yaml and source
Updates to placement of some glyphs; also kerning spacing and a test glyph to see what it will be like to extend this typeface to support other languages using this OT code. It looks like script the additions to the code is what makes best sense otherwise it will be a bit too labour intensive.